将配置文件检查移至init中
