+++ start.py
@@ -1,44 +1,21 @@
 from core import init,g_var
 from core.GUI import main
 import os
-import json
 
-init.init()
 
 def start():
     # 启动！
     g_var.GUI.MainWin = main.Main()
     g_var.GUI.MainWin.mainloop()
 
-# -----配置文件检查防止崩溃------
-def is_file_empty(file_path):
-    return os.stat(file_path).st_size == 0
-
-def restoration():
-    with open("./XCL/LoginData.json", "w") as file:
-        json.dump({
-            "status": False,
-            "token": None
-        }, file)
-    file.close()
-
-def detection():
-    try:
-        data = json.loads(open("./XCL/LoginData.json", "r").read())
-        a = data["status"]
-        b = data["token"]
-        data.close()
-        if a == "":
-            restoration()
-        if b == "":
-            restoration()
-    except:
-        restoration()
 
 # -----启动程序------
 if __name__ == '__main__':
-    if os.path.exists("./XCL/LoginData.json") and not is_file_empty("./XCL/LoginData.json"):
+    init.init()
+    if os.path.exists("./XCL/LoginData.json") and not init.is_file_empty("./XCL/LoginData.json"):
         start()
     else:
-        restoration()
-        start() # 补全文件后启动
+        # 配置文件检查防止崩溃
+        init.detection()
+        # 补全文件后启动
+        start()
